requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==5.0.1
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
azure-ai-anomalydetector==3.0.0b6
azure-cognitiveservices-anomalydetector==0.3.0
azure-common==1.1.28
azure-core==1.26.4
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==2.4.3
boto3==1.24.28
botocore==1.27.59
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
coverage==7.2.2
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.0
dask==2023.2.0
debugpy==1.5.1
decorator==5.1.1
distributed==2023.2.0
distro==1.8.0
exceptiongroup==1.0.4
executing==0.8.3
featuretools==0.22.0
fonttools==4.25.0
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.53.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
idna==3.4
imageio==2.27.0
importlib-metadata==6.0.0
importlib-resources==6.0.0
iniconfig==1.1.1
ipykernel==6.19.2
ipython==8.12.0
iso639==0.1.4
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.1.0
jupyter_core==5.3.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langdetect==1.0.9
lightfm==1.17
llvmlite==0.34.0
lmdb==1.4.1
locket==1.0.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mlblocks==0.4.1
mlprimitives==0.3.4
mock==5.0.2
msgpack==1.0.3
msrest==0.7.1
multidict==6.0.2
multitasking==0.0.11
munkres==1.1.4
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.8.1
numba==0.51.2
numexpr==2.8.4
numpy==1.20.3
oauthlib==3.2.2
opt-einsum==3.3.0
orion-ml==0.4.1
packaging==23.0
pandas==1.3.0
parso==0.8.3
partd==1.2.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
pip-autoremove==0.10.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pydot==1.4.2
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-louvain==0.13
pyts==0.10.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.1.0
referencing==0.30.0
regex==2023.5.4
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.7.2
s3fs==0.4.2
s3transfer==0.6.0
scikit-build==0.17.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setuptools==67.7.2
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.2.0
statsmodels==0.12.2
streamlit==1.25.0
tabulate==0.8.10
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.3.4
tensorflow-estimator==2.3.0
termcolor==2.3.0
threadpoolctl==2.2.0
tifffile==2023.4.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.3
wheel==0.40.0
wrapt==1.15.0
xgboost==0.90
XlsxWriter==1.3.9
yarl==1.8.1
yfinance==0.2.18
zict==2.2.0
zipp==3.11.0
